add edit remove update

+++ src/main/java/com/bmv/casestudycontact/repository/ContactRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ContactRepository extends CrudRepository<Contact, Integer> {
+public interface ContactRepository extends CrudRepository<Contact, Long> {
     List<Contact> findByNameContaining(String name);
 }
